requirements.txt
celery==4.4.0
-e git+https://github.com/czue/celery-progress@f56e19410a3be458de37e38e74f3a720cac5252f#egg=celery-progress
chembl-webresource-client==0.10.0
Django==2.2.8
django-celery-model==0.2.1
django-celery-results==1.1.2
django-polymorphic==2.1.2
djangorestframework==3.11.0
drf-yasg==1.17.0
# -e git+https://github.com/martin-sicho/DrugEx@fb4e4f89382966bbb55fa0cfb8f3fc82cc922833#egg=drugex
# jupyterlab==2.0.0b1
redis==3.3.11
requests==2.22.0
scikit-learn==0.22.1
pandas==0.22.0
seaborn==0.9.0
# torch==1.0.0
tqdm==4.41.1
molvs==0.1.1
psycopg2-binary==2.8.4
-e git+https://github.com/rdkit/django-rdkit.git@2565d4bf831bf43c997d81b3aa1a39842b7f56b1#egg=django-rdkit